% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.add_shrinkage.R
\name{ds.add_shrinkage}
\alias{ds.add_shrinkage}
\title{Shrink tree}
\usage{
ds.add_shrinkage(tree, shrinkage)
}
\arguments{
\item{tree}{The input tree.}

\item{shrinkage}{The factor by which the weights get multiplied.}
}
\value{
The changed tree
}
\description{
Shrink tree
}
