% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.calc_hist.R
\name{ds.calc_hist}
\alias{ds.calc_hist}
\title{Calculate the predicted output}
\usage{
ds.calc_hist(
  data_name,
  weight_update,
  amt_trees,
  dropout_rate,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{weight_update}{Through which method we choose the weights for our tree.}

\item{amt_trees}{How many trees have been built already.}

\item{dropout_rate}{Chance that a tree is not used for building the next
tree.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
WHich trees got removed from training.
}
\description{
Calculate the predicted output
}
