% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.calc_spsc.R
\name{ds.calc_spsc}
\alias{ds.calc_spsc}
\title{Calculate the split score and choose the best one}
\usage{
ds.calc_spsc(split_sums, spp_cand, reg_par)
}
\arguments{
\item{split_sums}{The histogram sums for all split bins.}

\item{spp_cand}{The split-point candidates for which the bins were
calculated.}

\item{reg_par}{The regularisation parameters.}
}
\value{
The split with the best split score.
}
\description{
Calculate the split score and choose the best one
}
