% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.create_data_split.R
\name{ds.create_data_split}
\alias{ds.create_data_split}
\title{Split the data into 'Test' and 'Training'}
\usage{
ds.create_data_split(
  data_name,
  train_test_ratio,
  split_status = NULL,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{train_test_ratio}{Percentage of the data which should be used for
Training.}

\item{split_status}{Defines if 'data_name' saves the full data, is already
split up into Training and Test data or saves only one of them.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
None.
}
\description{
Split the data into 'Test' and 'Training'
}
