% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.data_format_check.R
\name{ds.data_format_check}
\alias{ds.data_format_check}
\title{Check Data for basic rules}
\usage{
ds.data_format_check(data_name, datasources = NULL)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
NONE.
}
\description{
Check Data for basic rules
}
