% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exact_spp_cand.R
\name{ds.exact_spp_cand}
\alias{ds.exact_spp_cand}
\title{Creating exact Splitting-points}
\usage{
ds.exact_spp_cand(data_name, datasources = NULL)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The splitting point candidates.
}
\description{
Creating exact Splitting-points
}
