% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gen_factor_ord_spp_cand.R
\name{ds.gen_factor_spp_cand}
\alias{ds.gen_factor_spp_cand}
\title{Creation of the splitting-points for ordered factor features.}
\usage{
ds.gen_factor_spp_cand(amt_levels, amt_spp, selection_method)
}
\arguments{
\item{amt_levels}{The amount of levels this feature has.}

\item{amt_spp}{The amount of splitting-points which shall be created.}

\item{selection_method}{By which method the splitting-points shall be chosen.}
}
\value{
The created splitting-points.
}
\description{
Creation of the splitting-points for ordered factor features.
}
