% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gen_numeric_spp_cand.R
\name{ds.gen_numeric_spp_cand}
\alias{ds.gen_numeric_spp_cand}
\title{Creation of splitting-points for the numeric features}
\usage{
ds.gen_numeric_spp_cand(bounds, amt_spp, selection_method, add_par)
}
\arguments{
\item{bounds}{The bounds in which the splitting-points have to be.}

\item{amt_spp}{The amount of splitting-points which shall be created.}

\item{selection_method}{By which method the splitting-points shall be chosen.}

\item{add_par}{Additional parameters for the iterative hessian mode.}
}
\value{
The created splitting-points.
}
\description{
Creation of splitting-points for the numeric features
}
