% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gen_spp_cand.R
\name{ds.gen_spp_cand}
\alias{ds.gen_spp_cand}
\title{Generate Splitting point candidates}
\usage{
ds.gen_spp_cand(
  data_name,
  amt_spp,
  min_max,
  cand_select_mode = "percentile",
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{amt_spp}{How many splitting points per feature shall be created.}

\item{min_max}{What are the boundaries of each feature.}

\item{cand_select_mode}{Through which function shall the splitting points be
selected.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The created splitting points.
}
\description{
Generate Splitting point candidates
}
