% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.initialise_remote_tree.R
\name{ds.initialise_remote_tree}
\alias{ds.initialise_remote_tree}
\title{First training round of a remote tree.}
\usage{
ds.initialise_remote_tree(
  data_name,
  federation,
  weight_update,
  dropout_rate,
  cand_select,
  ithess_stop,
  split_method,
  reg_par,
  feature_subsampling,
  data_classes,
  amt_spp,
  max_splits,
  datasources
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{federation}{Through which method we share the data between client and
the servers.}

\item{weight_update}{Through which method we choose the weights for our tree.}

\item{dropout_rate}{Chance that a tree is not used for building the next
tree.}

\item{cand_select}{Splitting-point selection for numeric and factor features.}

\item{ithess_stop}{Maximum amount of times we update the split-point
candidates if the split-method is "totally_random".}

\item{split_method}{Through which method we choose the tree-splits.}

\item{reg_par}{Regularisation parameter which prevent overfitting.}

\item{feature_subsampling}{Which part of the feature space we use to build
the trees.}

\item{data_classes}{Data class for all features.}

\item{amt_spp}{The amount of splitting point candidates per feature.}

\item{max_splits}{The maximum amount of splits in the trained tree.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The trained Tree
}
\description{
First training round of a remote tree.
}
