% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ithess_spp_cand.R
\name{ds.ithess_spp_cand}
\alias{ds.ithess_spp_cand}
\title{Create Splitting-point candidates through the iterative hessian method.}
\usage{
ds.ithess_spp_cand()
}
\value{
The Splitting-point candidates
}
\description{
Create Splitting-point candidates through the iterative hessian method.
}
