% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.save_boosted_tree.R
\name{ds.save_boosted_tree}
\alias{ds.save_boosted_tree}
\title{Save the fully trained tree.}
\usage{
ds.save_boosted_tree()
}
\value{
The trained tree.
}
\description{
Save the fully trained tree.
}
