% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.save_variables.R
\name{ds.save_variables}
\alias{ds.save_variables}
\title{Save variables}
\usage{
ds.save_variables(data_name, variable_list, exist_check, datasources)
}
\arguments{
\item{data_name}{Name of the data.}

\item{variable_list}{Named list of variables which shall be saved.}

\item{exist_check}{If that variable should be checked for existance.}

\item{datasources}{Datasource connections.}
}
\value{
Nothing.
}
\description{
Save variables
}
