% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.summed_bins.R
\name{ds.summed_bins}
\alias{ds.summed_bins}
\title{Sum up bins over all split-point candidates.}
\usage{
ds.summed_bins(leaves, data_classes)
}
\arguments{
\item{leaves}{The histogram bins per leaf.}

\item{data_classes}{Data class for all features.}
}
\value{
The summed up bins.
}
\description{
Sum up bins over all split-point candidates.
}
