% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.train_boosted_tree.R
\name{ds.train_boosted_tree}
\alias{ds.train_boosted_tree}
\title{Training of a Gradient Boosted Decision Tree}
\usage{
ds.train_boosted_tree(
  data_name,
  train_test_ratio,
  split_status,
  max_treecount = 50,
  amt_spp,
  seed = NULL,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{train_test_ratio}{Percentage of the data which should be used for
Training.}

\item{split_status}{Defines if 'data_name' saves the full data, is already
split up into Training and Test data or saves only one of them.}

\item{max_treecount}{Maximum amount of trees to build our boosted decision
tree.}

\item{amt_spp}{The amount of split-points per feature.}

\item{seed}{If we want to choose a specific random behavior client side.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The trained decisiom tree model.
}
\description{
Training of a Gradient Boosted Decision Tree
}
