% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.uniform_spp_cand.R
\name{ds.uniform_spp_cand}
\alias{ds.uniform_spp_cand}
\title{Create Splitting-point candidates through uniform distribution}
\usage{
ds.uniform_spp_cand(amt_spp, min_max, seed = NULL)
}
\arguments{
\item{amt_spp}{The amount of splitting points per feature.}

\item{min_max}{The boundaries of every feature.}

\item{seed}{If wanted a seed for the random sample.}
}
\value{
The Splitting-point candidates.
}
\description{
Create Splitting-point candidates through uniform distribution
}
