% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.update_trees.R
\name{ds.update_trees}
\alias{ds.update_trees}
\title{Update the trees which got changed through DART.}
\usage{
ds.update_trees(data_name, removed_trees, last_tree_nmb, datasources)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{removed_trees}{Which trees got removed for training.}

\item{last_tree_nmb}{Number of the last trained tree.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
None.
}
\description{
Update the trees which got changed through DART.
}
