% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.update_weight.R
\name{ds.update_weight}
\alias{ds.update_weight}
\title{Update weights}
\usage{
ds.update_weight(
  data_name,
  current_tree,
  bounds_and_levels,
  max_splits,
  data_classes,
  reg_par,
  weight_update,
  loss_function,
  output_var,
  datasources
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{current_tree}{The fully chosen tree which misses its weights.}

\item{bounds_and_levels}{The maximum and minimum values for numeric features
and levels for factor features.}

\item{max_splits}{The maximum amount of splits in the trained tree.}

\item{data_classes}{List of data class per column.}

\item{reg_par}{Regularisation parameter which prevent overfitting.}

\item{weight_update}{Through which method we choose the weights for our tree.}

\item{loss_function}{The type of loss function under which we optimise the
tree.}

\item{output_var}{Name of the output variable.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The optimal weights for the current tree.
}
\description{
Update weights
}
