% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.calc_hist.R
\name{ds.calc_hist}
\alias{ds.calc_hist}
\title{Calculate the predicted output}
\usage{
ds.calc_hist(
  data_name,
  weight_update,
  last_tr_tree,
  amt_trees,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{weight_update}{Through which method we choose the weights for our tree.}

\item{last_tr_tree}{The last tree which was trained under the boosting step.}

\item{amt_trees}{How many trees have been built already.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
None.
}
\description{
Calculate the predicted output
}
