% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.calc_hist.R
\name{ds.calc_hist}
\alias{ds.calc_hist}
\title{Calculate the predicted output}
\usage{
ds.calc_hist(data_name, last_tr_tree, loss_function, datasources = NULL)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{last_tr_tree}{The last tree which was trained under the boosting step.}

\item{loss_function}{The loss function under which we optimize our boosted
tree building process.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
None.
}
\description{
Calculate the predicted output
}
