% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gen_spp_cand.R
\name{ds.gen_spp_cand}
\alias{ds.gen_spp_cand}
\title{Generate Splitting point candidates}
\usage{
ds.gen_spp_cand(
  data_name,
  cand_select,
  cand_hyper,
  data_classes,
  amt_spp,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{cand_select}{A data frame defining the split-point generation for all
different types.}

\item{cand_hyper}{Hyperparameter for all candidate selection methods.}

\item{data_classes}{List of data class per column.}

\item{amt_spp}{Amount of splitting points per feature.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The created splitting points.
}
\description{
Generate Splitting point candidates
}
