% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gen_spp_cand.R
\name{ds.gen_spp_cand}
\alias{ds.gen_spp_cand}
\title{Generate Splitting point candidates}
\usage{
ds.gen_spp_cand(bounds_and_levels, data_classes, cand_select, amt_spp)
}
\arguments{
\item{bounds_and_levels}{Bounds for numeric variables and level sets for
factor variables.}

\item{data_classes}{List of data class per column.}

\item{cand_select}{Which splitting-point candidate selection is used for
numeric and factor data.}

\item{amt_spp}{Amount of splitting points per feature.}
}
\value{
The created splitting points.
}
\description{
Generate Splitting point candidates
}
