% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.initialise_remote_tree.R
\name{ds.initialise_remote_tree}
\alias{ds.initialise_remote_tree}
\title{Title}
\usage{
ds.initialise_remote_tree(
  data_name,
  federation,
  weight_update,
  dropout_rate,
  cand_select,
  ithess_stop,
  split_method,
  reg_par,
  feature_subsampling,
  data_classes,
  amt_spp,
  max_splits,
  datasources
)
}
\arguments{
\item{data_name}{}

\item{federation}{}

\item{weight_update}{}

\item{dropout_rate}{}

\item{cand_select}{}

\item{ithess_stop}{}

\item{split_method}{}

\item{reg_par}{}

\item{feature_subsampling}{}

\item{data_classes}{}

\item{amt_spp}{}

\item{max_splits}{}

\item{datasources}{}
}
\value{
The trained Tree
}
\description{
Title
}
