% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.split_bins.R
\name{ds.split_bins}
\alias{ds.split_bins}
\title{Split the data into bins}
\usage{
ds.split_bins(
  data_name,
  min_max,
  spp_cand,
  current_tree,
  data_type,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{min_max}{The maximum and minimum values of all features.}

\item{spp_cand}{The candidates for a possible further split of the current
tree.}

\item{current_tree}{The tree which gets currently trained.}

\item{data_type}{Denotes for each feature if it is numeric or (originally)
categorical.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The histogram sums for each bin for all features.
}
\description{
Split the data into bins
}
