% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.train_boosted_tree.R
\name{ds.train_boosted_tree}
\alias{ds.train_boosted_tree}
\title{Training of a Gradient Boosted Decision Tree}
\usage{
ds.train_boosted_tree(
  data_name,
  train_test_ratio,
  max_treecount = 50,
  amt_spp,
  seed = NULL,
  drop_NA,
  bounds_and_levels,
  output_var,
  loss_function,
  cand_select,
  reg_par,
  spp_cand,
  drop_columns = NULL,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{train_test_ratio}{Percentage of the data which should be used for
Training.}

\item{max_treecount}{Maximum amount of trees to build our boosted decision
tree.}

\item{amt_spp}{The amount of split-points per feature.}

\item{seed}{If we want to choose a specific random behavior client side.}

\item{drop_NA}{If NA data in the output variable should be removed.}

\item{bounds_and_levels}{Bounds for numeric columns and levels for factors.}

\item{output_var}{The name of the column containing the output variable.}

\item{loss_function}{The name of the loss function we want to use for our
boosted tree.}

\item{cand_select}{Splitting-point selection for numeric and factor features.}

\item{reg_par}{Regularisation parameter which prevent overfitting.}

\item{spp_cand}{The splitting point candidates for each feature.}

\item{drop_columns}{Vector of data columns which shall be removed.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The trained decision tree model.
}
\description{
Training of a Gradient Boosted Decision Tree
}
