% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.train_tree.R
\name{ds.train_tree}
\alias{ds.train_tree}
\title{Train a single tree}
\usage{
ds.train_tree(
  data_name,
  last_tr_tree,
  loss_function,
  min_max,
  data_type,
  max_splits = 10,
  amt_spp,
  spp_cand = NULL,
  datasources = NULL
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{last_tr_tree}{The last trained tree.}

\item{loss_function}{The type of loss function under which we optimise the
tree.}

\item{min_max}{The maximum and minimum values of each feature space.}

\item{data_type}{Denotes for each feature if it is numeric or (originally)
categorical.}

\item{max_splits}{The maximum amount of splits in the trained tree.}

\item{amt_spp}{The amount of splitting point candidates per feature.}

\item{spp_cand}{The splitting point candidates for each feature.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The trained tree.
}
\description{
Train a single tree
}
