% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.update_weight.R
\name{ds.update_weight}
\alias{ds.update_weight}
\title{Update weights}
\usage{
ds.update_weight(
  data_name,
  current_tree,
  max_splits,
  reg_par,
  weight_update,
  datasources
)
}
\arguments{
\item{data_name}{The name under which the data is saved on the server.}

\item{current_tree}{The fully chosen tree which misses its weights.}

\item{max_splits}{The maximum amount of splits in the trained tree.}

\item{reg_par}{Regularisation parameter which prevent overfitting.}

\item{weight_update}{Through which method we choose the weights for our tree.}

\item{datasources}{DATASHIELD server connection.}
}
\value{
The optimal weights for the current tree.
}
\description{
Update weights
}
